Changes to the fix a rendering error in query builder - affected firefox and internet explorer, but not chrom or konqueror.
In query builder I changed the Model browser/Constraints table so that Contrain tile is aligned to the top and not floating in the middle vertically. Before the constrain tile moved out of view as the model browser tree was expanded in firefox and internet explorer. This was achieved by fusing the two cells on the right hand column into a single cell.

I also changed the Div CSS tag from " listHeading" to "heading" in constraintSetting.jsp so that the "Template contraint settings" heading used the same CSS as the other headings in the modelbrowser/constraints table.



Former-commit-id: 94cc9b3cbd31c6281be3d414c17c7ef9c0fb0049