Added a query xml shortening facility.
InterMine suffers from xml-itis. This badly affects GET requests with
long queries in them. This is another method for dealing with this issue
by allowing access to a minimal query storage service on the server,
which means we can refer to queries by id. At present these queries do
not live beyond the life-time of a session, as they live in memory. They
should be treated as very short-lived entities.


Former-commit-id: 4d8c0c9bebcd7e714a955f1b2a0d3e3a2a9c0282